+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Newest on top.
 
 ## [Unreleased]
 - The right-click add option is hidden for players the game has blocked (shown as "Unknown").
+- Players you've blocked in-game are now shown red in the party list and social window.
 
 ## [v0.0.0.243] - 2026-06-22
 - "Party Finder" can now only be selected while the global switch is on.